Move each command to its own module/file
Move each command that is currently defined into its own file/module that can
be accessed with `mod <command>`. This allows minimal imports for each file and
allows each file to be smaller and more self-contained.
